requirements.txt
# Django and related packages
Django==4.2.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
drf-yasg==1.21.7

# Command-line JSON processor
jq==1.7.0

# Language models and AI frameworks
langchain==0.1.16
langchain-community==0.0.32
langchain-openai==0.1.3
openai==1.55.3
langchain-text-splitters==0.0.1
chromadb==0.4.24
cassio==0.1.4
tiktoken==0.6.0
sentence-transformers==2.6.1
playwright
lxml
# Web scraping and parsing
beautifulsoup4==4.12.3

# Version control system
GitPython==3.1.43
jsonschema==4.23.0

#grpc
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.3

#trino
trino==0.329.0

#
lxml==5.3.0

# File processing
python-magic==0.4.27
chardet==5.2.0
CairoSVG==2.7.1
odfpy==1.4.1
nltk==3.9.1
python-pptx==1.0.2
unstructured==0.14.3